// resources/js/types/fitur-unggulan.d.ts
export interface FiturUnggulan {
    id: number
    icon: string
    title: string
    description: string
    image: string | null
    order: number
    is_active: boolean
    created_at: string
    updated_at: string
}

export interface FiturUnggulanFormData {
    icon: string
    title: string
    description: string
    order: number
    is_active: boolean
}